Update build_ffmpeg to work with current android_ndk
Toolchain and sysroot locations have changed within
//third_party/android_ndk. This change updates build_ffmpeg.py to use
the updated cross-compiler toolchains for android targets, and continues
to require the host to be a linux x64 machine. Notably, the new
android_ndk toolchains no longer include mips{64}el support, so that
portion of the (previously and currently) undocumented logic to attempt
to produce configs for an android mips* architecture is even less
supported now.

BUG=1170998

Change-Id: Idc20c608e07ff21dae8319c9beca3ac613cedd11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/third_party/ffmpeg/+/2713933
Reviewed-by: Ted Meyer <tmathmeyer@chromium.org>