Update build_ffmpeg to use /winsysroot instead of include paths
Before this change, cross-compilation targeting win-arm64 failed due
to the workaround for replacing white-space-containing include paths
with temporary symbolic links seems to have broken the include
dependencies. This change replaces all of the windows cross-compilation
include specifications in --extra-cflags with a new LLVM /winsysroot
option. Unfortunately, that /winsysroot option doesn't also replace the
need for specifying lib paths to lld-link; there is no /winsysroot
option for lld-link. So the previous workaround for whitespace-in-paths
is used for libpaths is kept.

Also included in this change is the removal of '--target=arm64-windows'
as an --extra-ldflags option to ffmpeg configure when cross-compiling to
that arch/os: lld-link does not recognize that option, even before this
change.

BUG=1170998

Change-Id: I123c5766ceb230a36ed1e818b76abb9220673a48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/third_party/ffmpeg/+/2737555
Reviewed-by: Ted Meyer <tmathmeyer@chromium.org>