Update build_ffmpeg's Android API level getter
Chromium, as of r858989, updated //build/config/android/config.gni to
use conditional logic to determine the Android API levels for each of 32
and 64 bit archs, breaking the simple regex search of that file
previously in build_ffmpeg. This change fixes that and hopefully
mitigates potential for similar future fragility by generating a full
temporary Chromium android GN config and querying it for the API levels.

BUG=1170998

Change-Id: I48d87aaf51aa7cd069d489ed059be8626f1ac4dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/third_party/ffmpeg/+/2740176
Reviewed-by: Ted Meyer <tmathmeyer@chromium.org>