Ensure libffmpeg.so is in the rpath of binaries
is_component_ffmpeg was introduced for the purpose of building FFmpeg as
a separate library that can be swapped out by Linux distros. However, at
some point since then, the functionality was broken, and it seems that
it went largely unnoticed due to many distros *also* unbundling FFmpeg.
As a result, Chromium's own FFmpeg libraries would never actually be
used, thus side-stepping any lookup issues.

This was discovered previously in https://crbug.com/1131061, with a fix
proposed, but the fix was Mac-specific, resulting in discussions as to
whether or not that build configuration is supported there in the first
place. By contrast, this CL only fixes it for Linux, as well as being
*significantly* smaller (and thus should not incur any major maintenance
overhead).

Bug: 1131061
Change-Id: Ic9bf3c0819979242b60dd53247dc140434711b61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/third_party/ffmpeg/+/3027653
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>