Check that no git add / rm autorename commands fail.
Previously, it forgot to && the results, so failures would go
unnoticed.  I had some un-removed autorename files, which i think
is either (a) an underlying problem that we should now be able to
debug if it happens again, or (b) my local repo was in a bad
state for some reason, but robosushi failed to notice.

Either way, checking the exit code of our git commands seems like
a really good idea.

Change-Id: If2d04ad47eeb9c0c571e5a15a541c65cd52f169c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/third_party/ffmpeg/+/3061500
Reviewed-by: Ted Meyer <tmathmeyer@chromium.org>